requirements.txt
aiohttp[speedups]
aiogram==3.20.0.post0
python-dotenv
validators